Maintain options while printing as DSL
This pipes a `MatchingOptions` instance throughout the print-as-DSL
methods, so that the right things can be printed as a consequence
of the matching options state. This would be nicer if PrettyPrinter
could just store the instance, but it's declared in `_RegexParser`,
so it doesn't have access to the `MatchingOptions` type.